PROBLEM 31 WORKING CONFIRMED!!! And it runs fast too :D
