+++ src/ProjectEuler.java
@@ -3,6 +3,7 @@ import java.io.FileReader;
 import java.io.IOException;
 import java.math.*;
 import java.util.ArrayList;
+import java.util.Arrays;
 import java.util.HashSet;
 
 /**
@@ -41,7 +42,7 @@ import java.util.HashSet;
  * @author Simon Alford, Manu Singhal
  *
  * Problems completed:
- *  1, 2, 4, 6, 7, 8, 10, 11, 13, 14, 16, 19, 20, 21, 22, 23, 25, 26, 27, 28, 29, 30
+ *  1, 2, 4, 6, 7, 8, 10, 11, 13, 14, 16, 19, 20, 21, 22, 23, 25, 26, 27, 28, 29, 30, 32, 33
  *  
  *  As always, contact Manu S.
  * (singhalmanu9 on GitHub) or visit the GitHub repository
@@ -956,6 +957,63 @@ public class ProjectEuler {
 	}
     
     /**
+	 * @author Simon Alford
+	 * Finds the number of ways to make 2 dollars with 2 dollar, 1 dollar, and 50, 25, 5, 2, and 1 cent coins.
+	 * I solved this problem recursively. I'm not sure if there's any other way to do it... It was definitely one
+	 * of the harder problems that I've solved (there were harder problems, but this one I actually solved by myself!!!).
+	 */
+	public static void problem31()
+	{
+		ArrayList<Integer> coins = new ArrayList<Integer>(Arrays.asList(200, 100, 50, 20, 10, 5, 2, 1));
+		System.out.println(numOfWays(coins, 200));
+	}
+		
+	/**
+	 * Recursive method for problem 31. Given a list of the coins to use, and the sum to make,
+	 * gives the number of ways to make the sum with the coins. Goes recursively:
+	 * @author Simon Alford
+	 * @param coins - an ArrayList of the coins (in cents) available to use to make the sum.
+	 * @param sum - the number of cents being made.
+	 * @return number of ways to make the sum with the ArrayList of coins.
+	 */
+	public static int numOfWays(ArrayList<Integer> coins, int sum)
+	{
+		/* the base case of the recursive function. If there is only one coin, or no coins
+		 * then there is only one way to make that sum (only one way to make 5c with pennies.)
+		 * I'm pretty sure this only works because the lowest denomination is pennies, so 
+		 * any number of cents will always be able to be made.
+		 */
+		if(coins.size() < 2)
+		{
+			//System.out.println("!");
+			return 1;
+		}
+		
+		int numOfWays = 0;
+		
+		//copies the coins into a new ArrayList so that it can be modified without
+		//modifying the original ArrayList.
+		ArrayList<Integer> rest = new ArrayList<Integer>(coins.size());
+		rest.addAll(coins);
+		
+		//Go through each coin in the list, removing each one.
+		for(int i = 0; i < coins.size() - 1; i++)
+		{
+			int coin = rest.remove(0);
+			//go through each number of the coin. Ex. a list of 5, 2, and 1 cent, needing to make
+			//10cents, you can have num be 1 or 2 to make 10 cents.
+			for(int num = 1; num*coin <= sum; num++)
+			{
+				//System.out.println(rest.toString());
+				//System.out.println(sum - num*coin + ", " + num + "*" + coin);
+				numOfWays+= numOfWays(rest, sum - num*coin);
+			}
+		}
+		// ++ to account for the last coin that wasn't gone through.
+		return ++numOfWays;
+	}
+    
+    /**
      * @author Simon Alford
      * This problem was pretty easy. Solved through brute force.
      */
@@ -993,9 +1051,24 @@ public class ProjectEuler {
 		System.out.println("Sum = " + sum);
 	}
     
+    /**
+     * Brute force solution.
+     * @author Simon Alford
+     */
+    public static void problem33()
+    {
+    	for(int i = 10; i < 100; i++)
+    	{
+    		for(int j = 10; j < 100; j++);
+    		{
+    			
+    		}
+    	}
+    }
+    
     
     public static void main(String[] args)
     {
-        problem11();
+        problem33();
     }    
 }
